refactored HTML conversions to new htmlutils, refactored dictionary conversion to model
